+++ python/tvm/tir/tensor_intrin/rocm.py
@@ -35,8 +35,8 @@ def sdot4(
         C[0] += T.call_llvm_pure_intrin(
             T.llvm_lookup_intrinsic_id("llvm.amdgcn.sdot4"),
             T.uint32(4),
-            T.reinterpret(A.vload([0], "int8x4"), dtype="int32"),
-            T.reinterpret(B.vload([0], "int8x4"), dtype="int32"),
+            T.reinterpret(A.vload([T.int32(0)], "int8x4"), dtype="int32"),
+            T.reinterpret(B.vload([T.int32(0)], "int8x4"), dtype="int32"),
             T.int32(0),
             T.bool(1),
             dtype="int32",
