+++ python/tvm/tir/tensor_intrin/x86.py
@@ -50,14 +50,14 @@ def dot_product_16x4_u8i8i32_vnni(
         T.reads(C[0:16], A[0:4], B[0:16, 0:4])
         T.writes(C[0:16])
 
-        A_u8x4 = A.vload([0], "uint8x4")
+        A_u8x4 = A.vload([T.int32(0)], "uint8x4")
         A_i32 = T.reinterpret(A_u8x4, dtype="int32")
 
-        B_i8x64 = B.vload([0, 0], dtype="int8x64")
+        B_i8x64 = B.vload([T.int32(0), T.int32(0)], dtype="int8x64")
         B_i32x16 = T.reinterpret(B_i8x64, dtype="int32x16")
-        C_i32x16 = C.vload([0], dtype="int32x16")
+        C_i32x16 = C.vload([T.int32(0)], dtype="int32x16")
 
-        C[T.ramp(T.int32(0), 1, 16)] = T.call_llvm_pure_intrin(
+        C[T.ramp(T.int32(0), T.int32(1), 16)] = T.call_llvm_pure_intrin(
             T.llvm_lookup_intrinsic_id("llvm.x86.avx512.vpdpbusd.512"),
             T.uint32(3),
             C_i32x16,
@@ -77,12 +77,12 @@ def dot_product_16x4_u8i8i32_avx512(
         T.reads(C[0:16], A[0:4], B[0:16, 0:4])
         T.writes(C[0:16])
 
-        A_u8x4 = A.vload([0], "uint8x4")
+        A_u8x4 = A.vload([T.int32(0)], "uint8x4")
         A_i32 = T.reinterpret(A_u8x4, dtype="int32")
         A_brdcst = T.broadcast(A_i32, 16)
         A_u8x64 = T.reinterpret(A_brdcst, dtype="uint8x64")
 
-        B_i8x64 = B.vload([0, 0], dtype="int8x64")
+        B_i8x64 = B.vload([T.int32(0), T.int32(0)], dtype="int8x64")
 
         Red = T.call_llvm_pure_intrin(
             T.llvm_lookup_intrinsic_id("llvm.x86.avx512.pmaddubs.w.512"),
@@ -92,7 +92,7 @@ def dot_product_16x4_u8i8i32_avx512(
             dtype="int16x32",
         )
 
-        C[T.ramp(T.int32(0), 1, 16)] += T.call_llvm_pure_intrin(
+        C[T.ramp(T.int32(0), T.int32(1), 16)] += T.call_llvm_pure_intrin(
             T.llvm_lookup_intrinsic_id("llvm.x86.avx512.pmaddw.d.512"),
             T.uint32(2),
             Red,
